+++ test/rules/lines-between-methods.js
@@ -3,7 +3,7 @@ const {errors, concat, extendToClassExpression} = require(`../helper`);
 module.exports = extendToClassExpression({
 	valid: [{
 		options: [{
-			newlinesCount: 2
+			amount: 2
 		}],
 		code: concat(
 			`class Klass1 {`,
@@ -22,7 +22,7 @@ module.exports = extendToClassExpression({
 		)
 	}, {
 		options: [{
-			newlinesCount: 2
+			amount: 2
 		}],
 		code: concat(
 			`class Klass1 {`,
@@ -54,7 +54,7 @@ module.exports = extendToClassExpression({
 	}],
 	invalid: [{
 		options: [{
-			newlinesCount: 2
+			amount: 2
 		}],
 		code: concat(
 			`class Klass {`,
@@ -72,11 +72,11 @@ module.exports = extendToClassExpression({
 			`}`
 		),
 		errors: errors(
-			`Count of newlines between methods should be 2, but 1 given.`
+			`Amount of newlines between methods should be 2, but 1 given.`
 		)
 	}, {
 		options: [{
-			newlinesCount: 2
+			amount: 2
 		}],
 		code: concat(
 			`class Klass {`,
@@ -106,7 +106,7 @@ module.exports = extendToClassExpression({
 			`}`
 		),
 		errors: errors(
-			`Count of newlines between methods should be 2, but 1 given.`
+			`Amount of newlines between methods should be 2, but 1 given.`
 		)
 	}]
 });