+++ test/rules/lines-between-props.js
@@ -3,7 +3,7 @@ const {errors, concat, extendToClassExpression} = require(`../helper`);
 module.exports = extendToClassExpression({
 	valid: [{
 		options: [{
-			newlinesCount: 2
+			amount: 2
 		}],
 		code: concat(
 			`class Klass1 {`,
@@ -51,7 +51,7 @@ module.exports = extendToClassExpression({
 	}],
 	invalid: [{
 		options: [{
-			newlinesCount: 2
+			amount: 2
 		}],
 		code: concat(
 			`class Klass {`,
@@ -96,8 +96,8 @@ module.exports = extendToClassExpression({
 			`}`
 		),
 		errors: errors(
-			`Count of newlines between props should be 2, but 1 given.`,
-			`Count of newlines between props should be 2, but 1 given.`
+			`Amount of newlines between props should be 2, but 1 given.`,
+			`Amount of newlines between props should be 2, but 1 given.`
 		)
 	}]
 });