+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change log
 
+## 0.2.0 (release date: 31.10.2017)
+
+* Migrate to ESLint 4
+* `newline-before-after-class`: param `newlinesCountBefore` renamed to `before` 
+* `newline-before-after-class`: param `newlinesCountAfter` renamed to `after` 
+* `newline-between-methods`: param `newlinesCount` renamed to `amount` 
+* `newline-between-props`: param `newlinesCount` renamed to `amount` 
+* `newline-between-statics`: param `newlinesCount` renamed to `amount` 
+* Renamed rules with `newline` prefix to be consistent with ESLint conventions:
+    - `newline-before-after-class` -> `lines-around-class`
+    - `newline-between-methods` -> `lines-between-methods`
+    - `newline-between-props` -> `lines-between-props`
+    - `newline-between-statics` -> `lines-between-statics`
+
 ## 0.1.2 (release date: 5.07.2017)
 
 * `space-in-typecast`: more robust implementation